Patch from Nick Bailey fixes various pitch tracker issues.
git-svn-id: http://svn.code.sf.net/p/rosegarden/code/trunk/rosegarden@12427 cf6a83e7-f017-4305-b0b3-678bde24f230

+++ src/sound/Tuning.cpp
@@ -579,7 +579,7 @@ double Tuning::getFrequency(Rosegarden::Pitch p) const {
     
     const int octaveDifference = octave - m_refOctave;
     
-    const int octaveRatio = pow( 2, octaveDifference );
+    const double octaveRatio = pow( 2, octaveDifference );
     
     ratio *= octaveRatio;
